Move ITs to the it package, moved unit tests to the root package
