+++ src/test/java/com/vaadin/flow/component/grid/GridNoSelectionModelTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.grid.tests;
+package com.vaadin.flow.component.grid;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
