+++ src/test/java/com/vaadin/flow/component/grid/it/GridViewIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.grid.tests;
+package com.vaadin.flow.component.grid.it;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -36,7 +36,7 @@ import com.vaadin.flow.demo.TabbedComponentDemoTest;
  * Integration tests for the {@link GridView}.
  *
  */
-public class GridIT extends TabbedComponentDemoTest {
+public class GridViewIT extends TabbedComponentDemoTest {
 
     @Test
     public void dataIsShown() throws InterruptedException {
